Add viewing of files in the commit where they are deleted

+++ usr/agite/view.adb
@@ -2,6 +2,7 @@ with Afpx, Temp_File;
 with Utils, Config;
 procedure View (Path : in String;
                 Hash : in Git_If.Git_Hash) is
+  Ok : Boolean;
 begin
   -- Create temporaty file
   declare
@@ -10,7 +11,11 @@ begin
     -- Cat file in it
     Afpx.Suspend;
     begin
-      Git_If.Cat (Path, Hash, Tmp_Name);
+      Ok := Git_If.Cat (Path, Hash, Tmp_Name);
+      if not Ok then
+        -- Try Hash^ (file might be deleted by the commit of hash)
+        Ok := Git_If.Cat (Path, Hash & "^", Tmp_Name);
+      end if;
       Afpx.Resume;
     exception
       when others =>
@@ -18,8 +23,10 @@ begin
         raise;
     end;
     -- Launch viewer
-    Utils.Launch (Config.Viewer & " " & Tmp_Name
-                & "; sleep 5; rm " & Tmp_Name);
+    if Ok then
+      Utils.Launch (Config.Viewer & " " & Tmp_Name
+                  & "; sleep 5; rm " & Tmp_Name);
+    end if;
   end;
     exception
       when Temp_File.Invalid_Dir | Temp_File.No_More_Temp =>
